Implementing D's Code
Included the ability to choose the starting values for the matrix from what Daniel has developed so far. Was able to get the display working with just Cell.h and Grid.h.  Commented the code describing what  I puled from Daniels code and what I stole from the internet in order to get this to work.

Current stage, spaghetti and meatballs.